improvements to chunk processor api
Prior to these changes, the chunk processor accepted a strbuf, and on
each iteration would populate it with the entire chunk from the png file.
This was not ideal since it needed to load the entire chunk into heap
memory which could (in theory) be significantly large. This also might
be less performant, since the buffer needs to be parsed anytime any of
the chunk fields are needed.

On each chunk_iterator_next(), the new implementation will parse the
length, type and CRC fields first, and then the caller can use
chunk_iterator_read_data to read into a fixed buffer from the chunk's
data.