use zlib compression library
Rather than write the input data plainly in the chunk's data section,
use DEFLATE compression provided by the zlib library.

The benefits of using zlib are two-fold. First, and most obvious, is
that the data is compressed before embedded in the file, which will help
bring the file size down. Secondly, this will obfuscate the message
enough that a quick byte inspection of the file won't immediately reveal
the message.

Rather than write the entire data in a single chunk, we write several
chunks with a data length of 8192.